requirements.txt
numpy>=1.24
torch>=2.0
gymnasium>=0.29
matplotlib>=3.7
imageio>=2.31
imageio-ffmpeg>=0.4
Pillow>=10.0
pyyaml>=6.0
pyserial>=3.5
